Rename test-utils "types" to "mocks"
